+++ src/CleanRegex/Internal/Match/GroupByFunction.php
@@ -1,7 +1,9 @@
 <?php
 namespace TRegx\CleanRegex\Internal\Match;
 
+use TRegx\CleanRegex\Exception\GroupNotMatchedException;
 use TRegx\CleanRegex\Exception\InvalidReturnValueException;
+use TRegx\CleanRegex\Internal\GroupKey\GroupKey;
 use TRegx\CleanRegex\Internal\Type\ValueType;
 use TRegx\CleanRegex\Match\Details\Detail;
 use TRegx\CleanRegex\Match\Details\Group\CapturingGroup;
@@ -26,8 +28,14 @@ class GroupByFunction
 
     private function groupKey($key): string
     {
-        if ($key instanceof Detail || $key instanceof CapturingGroup) {
+        if ($key instanceof Detail) {
             return $key->text();
+        }
+        if ($key instanceof CapturingGroup) {
+            if ($key->matched()) {
+                return $key->text();
+            }
+            throw GroupNotMatchedException::forGroupBy(GroupKey::of($key->usedIdentifier()));
         }
         if (\is_int($key) || \is_string($key)) {
             return $key;